+++ src/main/java/dev/aaronhowser/mods/geneticsresequenced/AntiFieldCarrier.java
@@ -1,0 +1,11 @@
+package dev.aaronhowser.mods.geneticsresequenced;
+
+import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
+
+public interface AntiFieldCarrier {
+
+	default LongOpenHashSet geneticsresequenced$getAntiFieldPositions() {
+		throw new IllegalStateException();
+	}
+
+}